ADD RequestBuilder, GiphyAPI and LoadTrendingListUseCase
The RequestBuilder is used to build a request, with options to set the http method, endpoint, parameters and headers. It returns a Requestable that can be used with Alamofire.

The GiphyAPI is a specific API for Giphy, with host, apiKey and decoder defined. It implements the request function from the RemoteAPI protocol.

The LoadTrendingListUseCase acts as glue between the GiphyAPI and the GiphyRepository, letting it fetch a list of trending gifs, and storing them in the GiphyRepository.